refactor!: move FieldAriaController to a11y-base, test with Lit
